feat: passed all tests

+++ src/albums/albums.service.ts
@@ -1,4 +1,10 @@
-import { Inject, Injectable } from '@nestjs/common';
+import {
+  forwardRef,
+  Inject,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
+import { FavoritesService } from 'src/favorites/favorites.service';
 import { TracksService } from 'src/tracks/tracks.service';
 import { v4 as uuidv4 } from 'uuid';
 import { CreateAlbumDto } from './dto/create-album.dto';
@@ -7,17 +13,24 @@ import { Album } from './interfaces/album.interface';
 
 @Injectable()
 export class AlbumsService {
-  private albums: Album[] = [];
+  constructor(
+    @Inject(forwardRef(() => TracksService))
+    private readonly tracksService: TracksService,
 
-  @Inject(TracksService)
-  private readonly tracksService: TracksService;
+    @Inject(forwardRef(() => FavoritesService))
+    private readonly favoritesService: FavoritesService,
+  ) {}
+
+  private albums: Album[] = [];
 
   async getAll(): Promise<Album[]> {
     return this.albums;
   }
 
   async getById(id: string): Promise<Album> {
-    return this.albums.find((album) => id === album.id);
+    const album = this.albums.find((album) => id === album.id);
+    if (album) return album;
+    throw new NotFoundException();
   }
 
   async create(albumDto: CreateAlbumDto): Promise<Album> {
@@ -30,21 +43,37 @@ export class AlbumsService {
   }
 
   async remove(id: string): Promise<Album> {
-    await this.tracksService.removeAlbums(id);
-    this.albums = this.albums.filter((album) => album.id !== id);
-    return;
+    const album = this.albums.find((album) => id === album.id);
+    if (album) {
+      await this.tracksService.removeAlbums(id);
+      await this.favoritesService.removeAlbum(id);
+      this.albums = this.albums.filter((album) => album.id !== id);
+      return;
+    }
+    throw new NotFoundException();
   }
 
   async update(id: string, albumDto: UpdateAlbumDto): Promise<Album> {
-    let updatedAlbum: Album | null = null;
+    const album = this.albums.find((album) => id === album.id);
+    if (album) {
+      let updatedAlbum: Album | null = null;
+      this.albums = this.albums.map((album) =>
+        album.id === id
+          ? (updatedAlbum = {
+              ...album,
+              ...albumDto,
+            })
+          : album,
+      );
+      return updatedAlbum;
+    }
+    throw new NotFoundException();
+  }
+
+  async removeArtist(id: string): Promise<void> {
     this.albums = this.albums.map((album) =>
-      album.id === id
-        ? (updatedAlbum = {
-            ...album,
-            ...albumDto,
-          })
-        : album,
+      album.artistId === id ? { ...album, artistId: null } : album,
     );
-    return updatedAlbum;
+    return;
   }
 }
