+++ src/artists/artists.service.ts
@@ -1,23 +1,40 @@
-import { Inject, Injectable } from '@nestjs/common';
+import {
+  forwardRef,
+  Inject,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
+import { AlbumsService } from 'src/albums/albums.service';
+import { FavoritesService } from 'src/favorites/favorites.service';
 import { v4 as uuidv4 } from 'uuid';
+import { TracksService } from './../tracks/tracks.service';
 import { CreateArtistDto } from './dto/create-artist.dto';
 import { UpdateArtistDto } from './dto/update-artist.dto';
 import { Artist } from './interfaces/artist.interface';
-import { TracksService } from './../tracks/tracks.service';
 
 @Injectable()
 export class ArtistsService {
-  private artists: Artist[] = [];
+  constructor(
+    @Inject(forwardRef(() => TracksService))
+    private readonly tracksService: TracksService,
 
-  @Inject(TracksService)
-  private readonly tracksService: TracksService;
+    @Inject(forwardRef(() => AlbumsService))
+    private readonly albumsService: AlbumsService,
+
+    @Inject(forwardRef(() => FavoritesService))
+    private readonly favoritesService: FavoritesService,
+  ) {}
+
+  private artists: Artist[] = [];
 
   async getAll(): Promise<Artist[]> {
     return this.artists;
   }
 
   async getById(id: string): Promise<Artist> {
-    return this.artists.find((artist) => id === artist.id);
+    const artist = this.artists.find((artist) => id === artist.id);
+    if (artist) return artist;
+    throw new NotFoundException();
   }
 
   async create(artistDto: CreateArtistDto): Promise<Artist> {
@@ -30,21 +47,31 @@ export class ArtistsService {
   }
 
   async remove(id: string): Promise<Artist> {
-    await this.tracksService.removeArtist(id);
-    this.artists = this.artists.filter((artist) => artist.id !== id);
-    return;
+    const artist = this.artists.find((artist) => id === artist.id);
+    if (artist) {
+      await this.tracksService.removeArtist(id);
+      await this.albumsService.removeArtist(id);
+      await this.favoritesService.removeArtist(id);
+      this.artists = this.artists.filter((artist) => artist.id !== id);
+      return;
+    }
+    throw new NotFoundException();
   }
 
   async update(id: string, artistDto: UpdateArtistDto): Promise<Artist> {
-    let updatedArtist: Artist | null = null;
-    this.artists = this.artists.map((artist) =>
-      artist.id === id
-        ? (updatedArtist = {
-            ...artist,
-            ...artistDto,
-          })
-        : artist,
-    );
-    return updatedArtist;
+    const artist = this.artists.find((artist) => id === artist.id);
+    if (artist) {
+      let updatedArtist: Artist | null = null;
+      this.artists = this.artists.map((artist) =>
+        artist.id === id
+          ? (updatedArtist = {
+              ...artist,
+              ...artistDto,
+            })
+          : artist,
+      );
+      return updatedArtist;
+    }
+    throw new NotFoundException();
   }
 }
